requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.31
alembic==1.7.4
databases==0.5.9
pydantic==1.8.2
uvicorn
sqlalchemy
PyYAML=6.0.1